+++ scripts/GW_backup
@@ -8,7 +8,7 @@ from astropy.time import Time
 
 from gototile.skymap import SkyMap
 
-from gtecs.alert.gcn import GCNNotice
+from gtecs.alert.notices import GCNNotice
 from gtecs.common import logging
 from gtecs.common.system import get_pid, make_pid_file
 from gtecs.obs import database as db
